Refactor API prototype
