+++ implementation/jobberknoll/api/deno.json
@@ -1,5 +1,8 @@
 {
   "name": "@jobberknoll/api",
   "version": "1.0.0",
-  "exports": "./mod.ts"
+  "exports": "./mod.ts",
+  "imports": {
+    "~/": "./"
+  }
 }
