+++ implementation/jobberknoll/api/mod.ts
@@ -1,2 +1,2 @@
-export { buildApp } from "./app.ts";
-export { IntController } from "./controllers/int-controller.ts";
+export { buildApp } from "~/app.ts";
+export { IntController } from "~/int/int-controller.ts";
