+++ implementation/jobberknoll/api/shared/openapi.ts
@@ -1,4 +1,4 @@
-export function jsonResponse<T>(schema: T, description: string) {
+export function jsonRes<T>(schema: T, description: string) {
   return {
     content: {
       "application/json": {
